Boards, Fix a few retarded warnings
git-svn-id: svn://svn.eoserv.net/eoserv/trunk@157 da1c5d5e-72c9-4536-8af9-af1c9134b0e6

+++ src/handlers/Init.cpp
@@ -49,7 +49,7 @@ CLIENT_F_FUNC(Init)
 		}
 		else
 		{
-			int mins_remaining = std::min(255.0, std::ceil(double(ban_expires - std::time(0)) / 60.0));
+			int mins_remaining = int(std::min(255.0, std::ceil(double(ban_expires - std::time(0)) / 60.0)));
 			reply.AddByte(INIT_BAN_TEMP);
 			reply.AddByte(mins_remaining);
 		}
